[CLI] fix loading of broken profile files
