+++ packages/cli/src/lib/errorHandler.ts
@@ -1,6 +1,7 @@
 /* eslint-disable no-console */
 
 import { ClientError } from "@scramjet/client-utils";
+import { displayFormat } from "../types";
 import { profileConfig } from "./config";
 import { displayError, displayObject } from "./output";
 
@@ -12,9 +13,25 @@ async function jsonize(err: ClientError) {
 
 const getExitCode = (_err: ClientError | Error) => 1;
 
+const getLogParams = () => {
+    let debug: boolean = false;
+    let format: displayFormat = "pretty";
+
+    try {
+        const { log: { format: profileFormat, debug: profileDebug } } = profileConfig.getConfig();
+
+        if (profileFormat) format = profileFormat;
+        if (profileDebug) debug = profileDebug;
+    } catch (error: any) {
+        displayError(error, debug);
+    }
+    return { format, debug };
+};
+
 export const errorHandler = async (err: Error) => {
     process.exitCode = getExitCode(err);
-    const { log: { format, debug } } = profileConfig.getConfig();
+
+    const { format, debug } = getLogParams();
 
     if (err instanceof ClientError) {
         if (debug)
@@ -30,4 +47,6 @@ export const errorHandler = async (err: Error) => {
         else displayError(err.message);
     } else
         displayError(err, debug);
+
+    process.exit(process.exitCode);
 };
